backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1

# PDF & doc processing
pypdf==5.1.0

# Vector store & embeddings (local, CPU)
chromadb==0.5.15
sentence-transformers==2.7.0

# PyTorch CPU only
torch==2.5.1
--extra-index-url https://download.pytorch.org/whl/cpu

# LangChain for RAG
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.5
langchain-groq==0.2.2
langchain-huggingface==0.1.2

# OpenAI SDK (streaming)
openai==1.55.0
groq==0.5.0

# Transformers for HuggingFace BART (CPU)
transformers==4.46.0
accelerate==1.1.1

# HTTP client
httpx==0.27.2
sse-starlette==2.1.3
